mental_health_journal/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
transformers
torch
sqlalchemy
psycopg2-binary